Store call stack addresses at pointer width
Call stack entries are instruction addresses but were stored in uint32
buffers. On the x64 client this truncated every address: win32
DebugHelp::getCallStack() narrowed StackWalk's DWORD64 AddrPC.Offset to 32
bits, so symbol lookup and crash reports were unusable.

Widen the storage, the getCallStack()/lookupAddress() signatures and the
printing to uint64. RenderWorld and MemoryManager keep their own call stack
buffers and are widened to match the new signature; MemoryManager's other
path already used uintptr_t.

This mirrors the same change on the server, where the identical code was
worse than truncation: passing a uint32 buffer to backtrace(), which writes
void*, overran every buffer by 2x. Shared files must stay in sync between
src and client-tools per README.md.

+++ src/engine/shared/library/sharedDebug/src/shared/CallStack.h
@@ -40,7 +40,8 @@ private:
 
 private:
 
-	uint32 m_callStack[S_callStack];
+	//-- Instruction addresses; must be pointer-width capable (see DebugHelp::getCallStack).
+	uint64 m_callStack[S_callStack];
 };
 
 // ======================================================================
